requirements.txt
altgraph==0.10.2
backports.shutil-get-terminal-size==1.0.0
bdist-mpkg==0.5.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
flake8==3.5.0
Flask==0.12.
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
macholib==1.5.1
MarkupSafe==1.0
mccabe==0.6.1
pathlib==1.0.1
pew==1.1.0
pipenv==8.3.2
psycopg2==2.7.3.2
py2app==0.7.3
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.5.3
pyparsing==2.0.1
python-dateutil==1.5
python-dotenv==0.7.1
pytz==2013.7
requests==2.18.4
shutilwhich==1.1.0
SQLAlchemy==1.1.14
twilio==6.7.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
web.py==0.40.dev0
Werkzeug==0.12.2
zope.interface==4.1.1
